+++ extensions/default/src/Toolbar/ToolbarLayoutSelector.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useCallback } from 'react';
+import React, { useEffect, useState, useCallback, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { LayoutSelector as OHIFLayoutSelector, ToolbarButton, LayoutPreset } from '@ohif/ui';
 
@@ -124,8 +124,9 @@ function ToolbarLayoutSelectorWithServices({
 }
 
 function LayoutSelector({
-  rows,
-  columns,
+  rows = 3,
+  columns = 4,
+  onLayoutChange = () => {},
   className,
   onSelection,
   onSelectionPreset,
@@ -134,22 +135,30 @@ function LayoutSelector({
   ...rest
 }: withAppTypes) {
   const [isOpen, setIsOpen] = useState(false);
+  const dropdownRef = useRef(null);
 
   const { customizationService } = servicesManager.services;
   const commonPresets = customizationService.get('commonPresets') || defaultCommonPresets;
   const advancedPresets =
     customizationService.get('advancedPresets') || generateAdvancedPresets({ servicesManager });
 
-  const closeOnOutsideClick = () => {
-    if (isOpen) {
+  const closeOnOutsideClick = event => {
+    if (isOpen && dropdownRef.current) {
       setIsOpen(false);
     }
   };
 
   useEffect(() => {
-    window.addEventListener('click', closeOnOutsideClick);
+    if (!isOpen) {
+      return;
+    }
+
+    setTimeout(() => {
+      window.addEventListener('click', closeOnOutsideClick);
+    }, 0);
     return () => {
       window.removeEventListener('click', closeOnOutsideClick);
+      dropdownRef.current = null;
     };
   }, [isOpen]);
 
@@ -169,7 +178,10 @@ function LayoutSelector({
       disableToolTip={tooltipDisabled}
       dropdownContent={
         DropdownContent !== null && (
-          <div className="flex ">
+          <div
+            className="flex"
+            ref={dropdownRef}
+          >
             <div className="bg-secondary-dark flex flex-col gap-2.5 p-2">
               <div className="text-aqua-pale text-xs">Common</div>
 
@@ -229,12 +241,6 @@ LayoutSelector.propTypes = {
   columns: PropTypes.number,
   onLayoutChange: PropTypes.func,
   servicesManager: PropTypes.object.isRequired,
-};
-
-LayoutSelector.defaultProps = {
-  columns: 4,
-  rows: 3,
-  onLayoutChange: () => {},
 };
 
 export default ToolbarLayoutSelectorWithServices;