+++ platform/ui/src/components/Table/Table.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
-const Table = ({ children, className, fullWidth, style }) => {
+const Table = ({ children, className = '', fullWidth = true, style = {} }) => {
   const classes = {
     base: 'text-lg text-white',
     fullWidth: {
@@ -19,12 +19,6 @@ const Table = ({ children, className, fullWidth, style }) => {
       {children}
     </div>
   );
-};
-
-Table.defaultProps = {
-  className: '',
-  fullWidth: true,
-  style: {},
 };
 
 Table.propTypes = {